+++ src/main/java/net/instantgratification/mcainclusive/model/CommonVillagerModelAccess.java
@@ -1,9 +1,0 @@
-// Copyright (C) 2026 Dasik (Rifaditya) | GNU GPLv3
-package net.instantgratification.mcainclusive.model;
-
-import net.minecraft.client.model.geom.ModelPart;
-
-public interface CommonVillagerModelAccess {
-    ModelPart getLeftBreastPart();
-    ModelPart getRightBreastPart();
-}